backend/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.0
django-cors-headers==4.9.0
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.5
djangorestframework_simplejwt==5.5.1
idna==3.11
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
urllib3==2.5.0
rapidfuzz==3.6.1
